src/budget-controller/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "budget-controller"
version = "0.1.0"
description = "Budget enforcement controller for Dataset Whisperer agents"
requires-python = ">=3.12"
dependencies = [
    "kubernetes==31.0.0",
    "httpx==0.28.1",
    "pydantic==2.10.3",
    "opentelemetry-api==1.28.4",
    "opentelemetry-sdk==1.28.4",
    "opentelemetry-exporter-otlp-proto-grpc==1.28.4",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.4",
    "pytest-asyncio==0.24.0",
    "respx==0.21.1",
    "ruff==0.8.6",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
